combine the WebUtils.convertMsToDate and CommandUtils.convertMsToDate to CommonUtils.convertMsToDate
