+++ core/server/common/src/main/java/alluxio/web/WebUtils.java
@@ -11,24 +11,14 @@
 
 package alluxio.web;
 
-import alluxio.Configuration;
 import alluxio.Constants;
-import alluxio.PropertyKey;
 
 import com.google.common.base.Preconditions;
-
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 
 import javax.annotation.concurrent.ThreadSafe;
 
 @ThreadSafe
 final class WebUtils {
-
-  private static final String DATE_FORMAT_PATTERN =
-      Configuration.get(PropertyKey.USER_DATE_FORMAT_PATTERN);
-
   /**
    * Converts a byte array to string.
    *
@@ -61,17 +51,6 @@ final class WebUtils {
 
     return String.format("%d day(s), %d hour(s), %d minute(s), and %d second(s)", days, hours,
         mins, secs);
-  }
-
-  /**
-   * Converts milliseconds to date.
-   *
-   * @param millis milliseconds
-   * @return input encoded as date
-   */
-  public static String convertMsToDate(long millis) {
-    DateFormat dateFormat = new SimpleDateFormat(DATE_FORMAT_PATTERN);
-    return dateFormat.format(new Date(millis));
   }
 
   /**
