+++ core/server/worker/src/main/java/alluxio/web/WebInterfaceWorkerGeneralServlet.java
@@ -13,6 +13,7 @@ package alluxio.web;
 
 import alluxio.RuntimeConstants;
 import alluxio.collections.Pair;
+import alluxio.util.CommonUtils;
 import alluxio.util.FormatUtils;
 import alluxio.worker.block.BlockStoreMeta;
 import alluxio.worker.block.BlockWorker;
@@ -111,7 +112,7 @@ public final class WebInterfaceWorkerGeneralServlet extends HttpServlet {
      * @return the start time
      */
     public String getStartTime() {
-      return WebUtils.convertMsToDate(mStartTimeMs);
+      return CommonUtils.convertMsToDate(mStartTimeMs);
     }
 
     /**
